Adding organize_input_files to lambda shared funcs
